refactor(protocol): extract group messaging into group_mesh module
Move all group messaging types, constants, methods, and tests from
protocol.rs into a dedicated group_mesh.rs module. Adds MAX_GROUP_MEMBERS
cap (256) and MAX_GROUP_CONTENT_LENGTH validation with test coverage.
